refactor(planner): extract SACADRL family from SocNav facade
